Create project main app
The main app of the project.

Currently an informative website only, which was created by "Nicepage",
showing: available actions for 2 different clients, the dog owners and
the daycare owners.

The stylesheets of the homepage is located in the "static" directory,
for better usage of resources in the future, with other application in
the project.

Signed-off-by: Ofir Matasas <ofir.matasas@gmail.com>