Creating dog owner homepage
When the dog owner is logged in to Hotails, or when the home is pressed
via navbar, All the daycares in the database will be shown to the dog
owner on its homepage, with the possibility to check their profile.
If the daycare doesn't have any profile picture, a default one will be
presented.

Signed-off-by: Ofir Matasas <ofir.matasas@gmail.com>
Signed-off-by: tamirmatok <tamirmatok123@gmail.com>