Add login and log-out functionallity.
Adding a landing page that is a login page to the Hotails website.
In order for a user to be able to access the site,
he must be a user and enter his details on this page.
After sucsessful log-in the application recognize the user with all of his details.